move tabs into <Tabs /> component
