--- plugins/parodos/src/components/ParodosPage/TabLabel.tsx
@@ -1,39 +1,0 @@
-import React, { forwardRef, type ReactNode } from 'react';
-import { makeStyles } from '@material-ui/core';
-import StarIcon from '@material-ui/icons/Star';
-import NotificationImportantIcon from '@material-ui/icons/NotificationImportant';
-
-const useStyles = makeStyles({
-  highlightedTab: {
-    position: 'absolute',
-    top: '1rem',
-    right: 0,
-    color: '#F70D1A',
-  },
-});
-
-export interface TabLabelProps {
-  icon: ReactNode;
-  highlighted: boolean;
-  children: ReactNode;
-}
-
-export const tabLabelCreator = ({
-  icon,
-  highlighted,
-}: {
-  icon: ReactNode;
-  highlighted: boolean;
-}) =>
-  forwardRef<HTMLSpanElement, TabLabelProps>(({ children, ...props }, ref) => {
-    const styles = useStyles();
-
-    return (
-      <span {...props} ref={ref}>
-        {icon}
-        {highlighted && <StarIcon className={styles.highlightedTab} />}
-        {children}
-        <NotificationImportantIcon />
-      </span>
-    );
-  });
